backend/pyproject.toml
[project]
name = "sundeck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py>=2.4.0",
    "flask>=3.1.3",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "rich>=14.3.3",
    "watchdog>=6.0.0",
    "pystray>=0.19.5",
    "requests>=2.33.1",
]

[dependency-groups]
dev = ["pyinstaller>=6.0"]
